Refactor: python, cts: Move StandbyTest into the _cts module.
No code changes.
